test(ui): split sidebar presentation e2e coverage
